Unfinished reading from local database, app not working properly.

+++ app/src/main/java/com/example/treadmillassistant/backend/workout/WorkoutCalendar.kt
@@ -25,7 +25,7 @@ class WorkoutCalendar(var currentDate: Date = Date(), var workoutList: MutableLi
         workout!!.mediaLink = newWorkout.mediaLink
     }
 
-    fun getWorkout(workoutID: Int): Workout? {
+    fun getWorkout(workoutID: Long): Workout? {
         return workoutList.find{ it.ID == workoutID}
     }
