+++ app/src/main/java/com/example/treadmillassistant/backend/workout/WorkoutPhase.kt
@@ -6,5 +6,5 @@ class WorkoutPhase(var duration: Int = 0,
                    var workoutPlanID: Int = -1,
                    var orderNumber: Int = -1,
                    val isFinished: Boolean = false,
-                   var ID: Int=-1) {
+                   var ID: Long =-1) {
 }
