+++ app/src/main/java/com/example/treadmillassistant/backend/workout/WorkoutPlan.kt
@@ -1,6 +1,6 @@
 package com.example.treadmillassistant.backend.workout
 
-class WorkoutPlan(var name: String = "default name", var workoutPhaseList: MutableList<WorkoutPhase> = mutableListOf<WorkoutPhase>(), var userID: Int = 0, var ID: Int = -1) {
+class WorkoutPlan(var name: String = "default name", var workoutPhaseList: MutableList<WorkoutPhase> = mutableListOf<WorkoutPhase>(), var userID: Long = 0, var ID: Long = -1) {
 
     fun addWorkoutPhaseList(workoutPhaseList: MutableList<WorkoutPhase>){
         this.workoutPhaseList = workoutPhaseList
@@ -14,7 +14,7 @@ class WorkoutPlan(var name: String = "default name", var workoutPhaseList: Mutab
         workoutPhaseList.remove(workoutPhase)
     }
 
-    fun getPhase(phaseID: Int): WorkoutPhase{
+    fun getPhase(phaseID: Long): WorkoutPhase{
         return workoutPhaseList.first {it.ID == phaseID}
     }
 }
